+++ src/factor_instance_provider/mod.rs
@@ -2,6 +2,7 @@ mod derivation_request;
 #[allow(clippy::module_inception)]
 mod factor_instance_provider;
 
+mod cache;
 mod gateway;
 mod next_derivation_peek_outcome;
 mod profile_extensions;
@@ -10,6 +11,7 @@ mod unfulfillable_request;
 mod unfulfillable_request_reason;
 mod unfulfillable_requests;
 
+pub use cache::*;
 pub use derivation_request::*;
 pub use factor_instance_provider::*;
 pub use gateway::*;
